# swagger-manual.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "Complete CRUD operations for managing contacts - CSE341 Project",
    "version": "1.0.0"
  },
  "host": "contacts-api-eb09.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/": {
      "get": {
        "summary": "API Information",
        "description": "Get basic API information",
        "responses": {
          "200": {
            "description": "API information",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "CSE 341 Contacts API"
                },
                "version": {
                  "type": "string",
                  "example": "1.0.0"
                }
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "summary": "Get all contacts",
        "description": "Retrieve all contacts from the database",
        "responses": {
          "200": {
            "description": "List of all contacts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "description": "Add a new contact to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateContact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "schema": {
              "$ref": "#/definitions/CreateContactResponse"
            }
          },
          "400": {
            "description": "Missing required fields",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Get contact by ID",
        "description": "Retrieve a specific contact by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Contact ID",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Invalid ID format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "summary": "Update contact by ID",
        "description": "Update an existing contact by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Contact ID",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateContact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid ID format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete contact by ID",
        "description": "Delete a contact from the database by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Contact ID",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid ID format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "507f1f77bcf86cd799439011"
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-01-01"
        }
      }
    },
    "CreateContact": {
      "type": "object",
      "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"],
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-01-01"
        }
      }
    },
    "CreateContactResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "507f1f77bcf86cd799439011"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Error message"
        }
      }
    },
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Operation completed successfully"
        }
      }
    }
  }
}